requirements.txt
# RAG from Scratch - 项目依赖

# API 客户端
openai>=1.0.0
python-dotenv>=1.0.0

# 向量检索
sentence-transformers>=2.2.0
numpy>=1.24.0
